LayoutTests/imported/w3c/web-platform-tests/storage-access-api/requestStorageAccess-dedicated-worker.sub.https.window-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: Unsupported permission name "storage-access".

Harness Error (TIMEOUT), message = null

TIMEOUT Workers don't inherit storage access Test timed out
NOTRUN Workers don't observe parent's storage access
